Shop intrigration
Shop is now in maze2 with gold needed to purchase and stats for players that update with item purchases

+++ Assets/Scripts/Inventory/EquipmentObject.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Default Equipment", menuName = "Inventory System/Items/Equipment")]
 public class EquipmentObject : ItemObject
 {
-    public float atkBonus;
-    public float defenseBonus;
-    public float stealthBonus;
+    public int atkBonus;
+    public int defenseBonus;
+    public int stealthBonus;
     public void Awake()
     {
         type = ItemType.Equipment;
